Remove mv2-manifest-instance e2e functionality
