+++ src/tests/end-to-end/common/browser-factory.ts
@@ -12,8 +12,6 @@ import {
     originalManifestCopyPath,
 } from 'tests/end-to-end/common/extension-paths';
 import { ManifestInstance } from 'tests/end-to-end/common/manifest-instance';
-import { ManifestV2Instance } from 'tests/end-to-end/common/mv2-manifest-instance';
-import { ManifestV3Instance } from 'tests/end-to-end/common/mv3-manifest-instance';
 import { testResourceServerConfigs } from '../setup/test-resource-server-config';
 import { Browser } from './browser';
 import { DEFAULT_BROWSER_LAUNCH_TIMEOUT_MS } from './timeouts';
@@ -35,23 +33,12 @@ export interface ExtensionOptions {
 export async function launchBrowser(extensionOptions: ExtensionOptions): Promise<Browser> {
     const browserInstanceId = generateUID();
 
-    const manifestV3 = process.env.WEB_E2E_TARGET?.includes('mv3');
-    let manifestInstance: ManifestInstance;
-    let onCloseCallback: () => Promise<void>;
-    if (manifestV3) {
-        const originalManifestContent = await ManifestV3Instance.parse(originalManifestCopyPath);
-        manifestInstance = new ManifestV3Instance(originalManifestContent);
+    const originalManifestContent = await ManifestInstance.parse(originalManifestCopyPath);
+    const manifestInstance = new ManifestInstance(originalManifestContent);
 
-        onCloseCallback = async () => {
-            await new ManifestV3Instance(originalManifestContent).writeTo(manifestPath);
-        };
-    } else {
-        const originalManifestContent = await ManifestV2Instance.parse(originalManifestCopyPath);
-        manifestInstance = new ManifestV2Instance(originalManifestContent);
-        onCloseCallback = async () => {
-            await new ManifestV2Instance(originalManifestContent).writeTo(manifestPath);
-        };
-    }
+    const onCloseCallback = async () => {
+        await new ManifestInstance(originalManifestContent).writeTo(manifestPath);
+    };
 
     const manifestCopy = AddExtraManifestPermissions(
         extensionOptions.addExtraPermissionsToManifest,
@@ -70,7 +57,7 @@ export async function launchBrowser(extensionOptions: ExtensionOptions): Promise
     // simplify the initial migration from Puppeteer to Playwright.
     const playwrightContext = await launchNewBrowserContext(userDataDir, extensionPath);
 
-    const browser = new Browser(browserInstanceId, playwrightContext, manifestV3, onCloseCallback);
+    const browser = new Browser(browserInstanceId, playwrightContext, true, onCloseCallback);
 
     const backgroundPage = await browser.background();
     if (extensionOptions.suppressFirstTimeDialog) {
